+++ app/src/main/java/io/webthings/app/AppScaffold.kt~f80d7effdfe277ac8f83fa6f1c45b1ee9ed7bc61
@@ -1,19 +1,20 @@
-package io.webthings.webthings
+package io.webthings.app
 
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.rememberNavController
-import io.webthings.webthings.ui.menus.Drawer
-import io.webthings.webthings.ui.theme.WebthingsTheme
-import io.webthings.webthings.utils.MainViewModel
-import io.webthings.webthings.utils.NavigationHost
+import io.webthings.app.ui.menus.Drawer
+import io.webthings.app.ui.theme.HeaderStyle
+import io.webthings.app.ui.theme.WebthingsTheme
+import io.webthings.app.utils.BackPressHandler
+import io.webthings.app.utils.MainViewModel
+import io.webthings.app.utils.NavigationHost
 import kotlinx.coroutines.launch
 
 @Composable
@@ -24,9 +25,17 @@ fun AppScaffold(){
     val scope = rememberCoroutineScope()
     val currentScreen by viewModel.currentScreen.observeAsState()
 
+    if (scaffoldState.drawerState.isOpen){
+        BackPressHandler {
+            scope.launch {
+                scaffoldState.drawerState.close()
+            }
+        }
+    }
+
     val topBar: @Composable () -> Unit= {
         TopAppBar(
-            title = { Text(currentScreen!!.title) },
+            title = { Text(currentScreen!!.title, style = HeaderStyle) },
             navigationIcon = {
                 //TODO if not home screen change to back arrow
                 IconButton(onClick = {
