+++ app/src/main/java/io/webthings/app/ui/NavRoutes.kt
@@ -1,37 +1,21 @@
-package io.webthings.app.ui
-
+package io.webthings.webthings.ui
 
 import android.content.Context
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Face
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.ui.graphics.vector.ImageVector
-import io.webthings.app.R
 import android.content.res.Resources
-
-private val _context = Resources.getSystem()
+import androidx.compose.ui.res.stringResource
+import io.webthings.webthings.R
+import kotlin.math.absoluteValue
 
 sealed class NavRoutes(val route:String, val title:String){
-
     object HomeScreen : NavRoutes("home",
-        _context.getString(R.string.home))
-
-    sealed class DrawerItems(
-        route: String,
-        title: String,
-        val icon: ImageVector): NavRoutes(route, title){
-        object DashboardScreen: DrawerItems("Dashboard",
-            "Main Dashboard",
-            Icons.Default.Face)
-            object GwSettingsScreen: DrawerItems("GwSettings",
-                "Settings",
-                Icons.Default.Settings)
-        }
-
+        "home")
+    object DashboardScreen: NavRoutes("Dashboard",
+        "Main Dashboard")
+    object GwSettingsScreen: NavRoutes("GwSettings", "Settings")
 }
 
 val drawerMenuItems = listOf(
     //TODO add other drawer menu entry
-    NavRoutes.DrawerItems.DashboardScreen,
-    NavRoutes.DrawerItems.GwSettingsScreen
+    NavRoutes.DashboardScreen,
+    NavRoutes.GwSettingsScreen
 )
