--- app/src/main/java/io/webthings/webthings/ui/theme/Color.kt
@@ -1,15 +1,0 @@
-package io.webthings.webthings.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple200 = Color(0xFFBB86FC)
-val Purple500 = Color(0xFF6200EE)
-val Purple700 = Color(0xFF3700B3)
-val Teal200 = Color(0xFF03DAC5)
-
-//Webthings Color
-
-val LightBlue = Color(93,155,199)
-val DarkBlue = Color(72, 120, 154)
-val White = Color(255,255,255)
-val DarkWhite = Color(221,221,221)
